+++ modules/hanna-react/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Add default Polish translation for `Layout`
 - feat: Make prop `title` optional for `MainMenu`
 
 ## 0.10.96